+++ src/client/datascience/jupyter/kernels/types.ts
@@ -6,11 +6,11 @@
 import type { KernelMessage, Session } from '@jupyterlab/services';
 import type { Observable } from 'rxjs/Observable';
 import type { CancellationToken, Event, QuickPickItem, Uri } from 'vscode';
+import { NotebookCell, NotebookDocument } from '../../../../../types/vscode-proposed';
 import type { ServerStatus } from '../../../../datascience-ui/interactive-common/mainState';
 import type { IAsyncDisposable, Resource } from '../../../common/types';
 import type { PythonInterpreter } from '../../../pythonEnvironments/info';
 import type {
-    ICell,
     IJupyterKernel,
     IJupyterKernelSpec,
     IJupyterSessionManager,
@@ -65,9 +65,10 @@ export interface IKernel extends IAsyncDisposable {
     readonly disposed: boolean;
     readonly kernelSocket: Observable<KernelSocketInformation | undefined>;
     start(): Promise<void>;
-    interrupt(timeoutInMs: number): Promise<InterruptResult>;
-    restart(timeoutInMs: number): Promise<void>;
-    executeObservable(code: string, file: string, line: number, id: string, silent: boolean): Observable<ICell[]>;
+    interrupt(): Promise<InterruptResult>;
+    restart(): Promise<void>;
+    executeCell(cell: NotebookCell): Promise<void>;
+    executeAllCells(document: NotebookDocument): Promise<void>;
     registerIOPubListener(listener: (msg: KernelMessage.IIOPubMessage, requestId: string) => void): void;
 }
