+++ src/client/datascience/notebook/serviceRegistry.ts
@@ -7,17 +7,15 @@ import { IExtensionSingleActivationService } from '../../activation/types';
 import { IServiceManager } from '../../ioc/types';
 import { KernelProvider } from '../jupyter/kernels/kernelProvider';
 import { NotebookContentProvider } from './contentProvider';
-import { NotebookExecutionService } from './executionService';
 import { NotebookIntegration } from './integration';
 import { VSCodeKernelPickerProvider } from './kernelProvider';
 import { NotebookDisposeService } from './notebookDisposeService';
 import { NotebookOutputRenderer } from './renderer';
 import { NotebookSurveyBanner, NotebookSurveyDataLogger } from './survey';
-import { INotebookContentProvider, INotebookExecutionService } from './types';
+import { INotebookContentProvider } from './types';
 
 export function registerTypes(serviceManager: IServiceManager) {
     serviceManager.addSingleton<INotebookContentProvider>(INotebookContentProvider, NotebookContentProvider);
-    serviceManager.addSingleton<INotebookExecutionService>(INotebookExecutionService, NotebookExecutionService);
     serviceManager.addSingleton<IExtensionSingleActivationService>(
         IExtensionSingleActivationService,
         NotebookIntegration
