+++ src/client/datascience/notebook/types.ts
@@ -3,19 +3,7 @@
 
 'use strict';
 
-import type { CancellationToken } from 'vscode';
-import type {
-    NotebookCell,
-    NotebookContentProvider as VSCodeNotebookContentProvider,
-    NotebookDocument
-} from 'vscode-proposed';
-
-export const INotebookExecutionService = Symbol('INotebookExecutionService');
-export interface INotebookExecutionService {
-    cancelPendingExecutions(document: NotebookDocument): void;
-    executeCell(document: NotebookDocument, cell: NotebookCell, token: CancellationToken): Promise<void>;
-    executeAllCells(document: NotebookDocument, token: CancellationToken): Promise<void>;
-}
+import type { NotebookContentProvider as VSCodeNotebookContentProvider, NotebookDocument } from 'vscode-proposed';
 
 export const INotebookContentProvider = Symbol('INotebookContentProvider');
 export interface INotebookContentProvider extends VSCodeNotebookContentProvider {
